# social-mini-app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Intelligence Platform API",
    "description": "Public REST API for the Social Intelligence Platform. All endpoints require a valid API key. Internal keys have higher rate limits and (in Sprint 2) access to write endpoints; external keys are read-only and rate-limited.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/admin/keys": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create a new API key",
        "description": "Create a new API key. Returns the raw key exactly once in the response - the server only retains the hash. Requires the admin token in the `Authorization: Bearer <ADMIN_TOKEN>` header. This endpoint does not require or use a regular X-API-Key.",
        "operationId": "create_api_key_v1_admin_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Key created. The `api_key` field in `data` is the raw key, shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_CreateKeyData_"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin token"
          },
          "403": {
            "description": "Invalid admin token"
          },
          "422": {
            "description": "Invalid request body"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AdminAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "APIKeyTier": {
        "type": "string",
        "enum": [
          "internal",
          "external_default",
          "external_elevated"
        ],
        "title": "APIKeyTier",
        "description": "API key tier. Controls rate limits and (in Sprint 2) write access."
      },
...
      "CreateKeyData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tier": {
            "$ref": "#/components/schemas/APIKeyTier"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tier",
          "key_prefix",
          "created_at",
          "api_key"
        ],
        "title": "CreateKeyData",
        "description": "Response body for a newly-created API key.\n\nThe `api_key` field is the only place the raw key will ever appear.\nSave it immediately - the server cannot recover it later."
      },
      "CreateKeyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Human-readable label for the key (e.g. 'Mini App Prod')."
          },
          "tier": {
            "$ref": "#/components/schemas/APIKeyTier",
            "description": "Tier: `internal`, `external_default`, or `external_elevated`."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "tier"
        ],
        "title": "CreateKeyRequest",
        "description": "Request body for creating a new API key."
      },
...
      "ResponseEnvelope_CreateKeyData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateKeyData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseEnvelope[CreateKeyData]"
      },
...
      "ResponseMeta": {
        "properties": {
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "ResponseMeta",
        "description": "Standard response metadata.\n\nAlways present in success responses. Fields are optional so the\nsame shape works for single-resource and paginated list responses."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin token from ADMIN_TOKEN env var.",
        "scheme": "bearer"
      },
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued via the admin endpoint. Internal keys have higher rate limits; external keys are read-only. The raw key is shown exactly once at creation time."
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin token from the ADMIN_TOKEN env var. Used only by the /v1/admin/* endpoints. A compromised regular API key cannot be used to mint more keys."
      }
    }
  },
  "security": [
    {
      "APIKeyAuth": []
    }
  ]
}